caas-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-cloud-firestore==2.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
python-dotenv==1.0.0
langchain==0.0.350
langchain-google-vertexai==0.0.3
anthropic==0.7.8
google-auth==2.23.4
google-api-python-client==2.108.0
firebase-admin==6.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2